app/requirements.txt
fastapi==0.115.6
passlib==1.7.4
psycopg2-binary==2.9.10
python-jose==3.3.0
uvicorn==0.33.0